- Implemented GSManager - Improved PadManager - Improved loading malloc_page_size from ELF64 - Emulated more PPU instructions/syscalls - Improved memory allocation

+++ trunk/Utilites/IdManager.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Array.h"
 
-typedef u8 ID_TYPE;
+typedef u32 ID_TYPE;
 
 struct ID
 {
